requirements.txt
Django==4.0.1
face_recognition==1.3.0
imutils==0.5.4
keras==3.3.3
mediapipe==0.10.14
mysql_connector_repackaged==0.3.1
numpy==1.24.3
Pillow==9.4.0
Pillow==10.3.0
pyttsx3==2.90
scipy==1.11.1
tensorflow==2.16.1
